azure-sageinsure-complete/fnol-processor/backend/move_folders/__init__.py
import json
import logging
from azure.functions import HttpRequest, HttpResponse
from azure.storage.blob import BlobServiceClient
import os

def main(req: HttpRequest) -> HttpResponse:
    try:
        req_body = req.get_json()
        source_container = req_body.get('source_container', 'scanning-staging')
        destination_container = req_body.get('destination_container', 'scanning-in-process')
        folder_key = req_body.get('folder_key', '')
        
        # Simulate folder move operation
        logging.info(f"Moving folder {folder_key} from {source_container} to {destination_container}")
        
        return HttpResponse(
            json.dumps({
                'statusCode': 200,
                'prefix': folder_key,
                'body': 'Folder moved successfully'
            }),
            status_code=200,
            mimetype="application/json"
        )
    except Exception as e:
        return HttpResponse(json.dumps({"error": str(e)}), status_code=500)