azure-sageinsure-complete/fnol-processor/backend/requirements.txt
azure-functions
azure-identity
azure-storage-blob
azure-ai-formrecognizer
azure-cosmos
openai
requests
python-dotenv
PyPDF2